Conclusion added.
Paper done, just needs some review before the push. The reason the
commits are all within the span of a couple days is because I prefer to
write papers in one sitting, or as close to that as possible. If I
spread it out to much I tend to lose my train of thought.